swarmstar/pyproject.toml
[tool.poetry]
name = "swarmstar"
version = "0.0.7"
description = "chatgpt on crack"
authors = ["BrianP8701 <brianp8701@gmail.com>"]
readme = "README.md"
include = ["swarmstar/internal_metadata/action_space.bson", "swarmstar/internal_metadata/action_space_metadata.json"]
documentation = "https://brianp8701.github.io/swarmstar/"

[tool.poetry.dependencies]
python = "^3.10"
importlib = "^1.0.4"
pydantic = "^2.5.3"
openai = "^1.9.0"
instructor = "^1.3.4"
pymongo = "^4.6.1"
docker = "^7.0.0"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
pyodbc = "^5.1.0"
pygithub = "^2.3.0"
aiosqlite = "^0.20.0"
greenlet = "^3.0.3"
strawberry-graphql = "^0.237.3"
python-dotenv = "^1.0.1"
stripe = "^10.8.0"
svix = "^1.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
python-dotenv = "^1.0.1"
coverage = "^7.4.2"
isort = "^5.13.2"

[tool.poetry.group.docs.dependencies]
mkdocs-material = {extras = ["imaging"], version = "^9.5.9"}
mkdocs = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
